# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOA Prototype API",
    "version": "0.1.0",
    "description": "Demo authentication and payments only. Not a production PG or customs API."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/requests": {
      "post": {
        "summary": "비공개 결제 대기 부탁 생성 (PAYMENT_PENDING)",
        "operationId": "post_requests",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRequest"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          }
        }
      }
    },
...
    "/requests/{id}/offers": {
      "post": {
        "summary": "단건 제안",
        "operationId": "post_requests_id_offers",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelerOffer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferInput"
              }
            }
          }
        }
      }
    },
    "/bundles/offers": {
      "post": {
        "summary": "동일 장소 일괄 제안",
        "operationId": "post_bundles_offers",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleOfferInput"
              }
            }
          }
        }
      }
    },
...
    "/requests/{id}/pay": {
      "post": {
        "summary": "구매자 선결제 후 부탁 공개",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8, "maxLength": 100 } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["expectedRevision", "paymentMethod", "paymentReference"], "properties": { "expectedRevision": { "type": "integer", "minimum": 0 }, "paymentMethod": { "type": "string", "enum": ["CARD", "ACCOUNT"] }, "paymentReference": { "type": "string", "minLength": 4, "maxLength": 40 }, "simulateFailure": { "type": "boolean", "default": false } } } } } },
        "responses": { "201": { "description": "한 번만 모의 보관 후 공개된 부탁", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } } } }, "409": { "description": "상태 충돌, 권한 불일치, 이미 결제됨 또는 모의 승인 실패" } }
      }
    },
    "/requests/{id}/cancel": {
      "post": {
        "summary": "여행자 선택 전 부탁 취소·모의 결제 전액 환불",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8, "maxLength": 100 } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["expectedRevision"], "properties": { "expectedRevision": { "type": "integer", "minimum": 0 } } } } } },
        "responses": { "201": { "description": "취소된 부탁", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } } } }, "409": { "description": "이미 선택됨 또는 revision/권한 충돌" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, random 24-hour demo session token"
      }
    },
    "schemas": {
...
      "CreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "productUrl": {
            "type": "string"
          },
          "productImage": {
            "type": "string",
            "maxLength": 2800000,
            "description": "Image data URL or empty string when receiptImage is provided"
          },
          "art": {
            "type": "string",
            "enum": [
              "keyring",
              "plush",
              "pouch",
              "tshirt",
              "pin",
              "bag"
            ]
          },
          "placeId": {
            "type": "string"
          },
          "localPrice": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2000000
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "desiredDate": {
            "type": "string",
            "format": "date"
          },
          "deliveryCountry": {
            "type": "string",
            "enum": [
              "KR",
              "JP"
            ]
          },
          "deliveryCity": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "CHARACTER",
              "GAME",
              "POPUP",
              "LOCAL",
              "FASHION",
              "CONCERT"
            ]
          },
          "option": {
            "type": "string"
          },
          "transport": {
            "type": "string",
            "enum": [
              "INTERNATIONAL_SHIPPING",
              "DOMESTIC_PARCEL",
              "MEETUP"
            ]
          }
        },
        "required": [
          "productName",
          "placeId",
          "localPrice",
          "quantity",
          "desiredDate",
          "category",
          "transport"
        ]
      },
...
      "OfferInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tripId": {
            "type": "string"
          },
          "reward": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "estimatedPurchaseDate": {
            "type": "string",
            "format": "date"
          },
          "estimatedDeliveryDate": {
            "type": "string",
            "format": "date"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "transport": {
            "type": "string",
            "enum": [
              "DOMESTIC_PARCEL",
              "MEETUP"
            ]
          }
        },
        "required": [
          "tripId",
          "reward",
          "estimatedPurchaseDate",
          "estimatedDeliveryDate",
          "message",
          "transport"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "ProductRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "productName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "productUrl": {
            "type": "string"
          },
          "productImage": {
            "type": "string",
            "maxLength": 2800000,
            "description": "image/jpeg, image/png or image/webp base64 data URL; no remote fetch"
          },
          "art": {
            "type": "string",
            "enum": [
              "keyring",
              "plush",
              "pouch",
              "tshirt",
              "pin",
              "bag"
            ]
          },
          "placeId": {
            "type": "string"
          },
          "localPrice": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2000000
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "desiredDate": {
            "type": "string",
            "format": "date"
          },
          "deliveryCountry": {
            "type": "string",
            "enum": [
              "KR",
              "JP"
            ]
          },
          "deliveryCity": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "CHARACTER",
              "GAME",
              "POPUP",
              "LOCAL",
              "FASHION",
              "CONCERT"
            ]
          },
          "option": {
            "type": "string"
          },
          "transport": {
            "type": "string",
            "enum": [
              "INTERNATIONAL_SHIPPING",
              "DOMESTIC_PARCEL",
              "MEETUP"
            ]
          },
          "requesterId": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "enum": [
              "KR",
              "JP"
            ]
          },
          "city": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "enum": [
              "JPY",
              "KRW"
            ]
          },
          "storeName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PAYMENT_PENDING",
              "REQUESTED",
              "OFFER_RECEIVED",
              "MATCHED",
              "PAYMENT_HELD",
              "PURCHASED",
              "TRAVELING",
              "SHIPPED",
              "DELIVERED",
              "CONFIRMED",
              "SETTLED",
              "CANCELLED",
              "DISPUTED"
            ]
          },
          "revision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "directPurchase": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "OFFLINE_ONLY"
            ]
          }
        }
      },
      "TravelerOffer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "tripId": {
            "type": "string"
          },
          "reward": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 100000
          },
          "estimatedPurchaseDate": {
            "type": "string",
            "format": "date"
          },
          "estimatedDeliveryDate": {
            "type": "string",
            "format": "date"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "transport": {
            "type": "string",
            "enum": [
              "INTERNATIONAL_SHIPPING",
              "DOMESTIC_PARCEL",
              "MEETUP"
            ]
          },
          "travelerId": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "bundleId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "CANCELLED"
            ]
          }
        }
      },
...
    }
  }
}